Move codeActions code to new file
